refactor(widget): one ConversationProvider + drop dead taskProgress/setState/TabBar
The SSE reducer already models conversation as one SseState{messages,task}, but
state was split across ChatContext+TaskContext (each with one consumer), forcing
_setMessages/messagesRef/taskRef injection so commit could hand-marshal one
transition across two stores — and the two setState calls could TEAR (messages
update while task lags post-await).

- Merge ChatContext+TaskContext into one ConversationProvider holding
  {messages,task}; the SSE effect commits every transition through a single
  atomic setState(prev=>transition(prev)). Delete _setMessages, messagesRef,
  taskRef, useChatContextInternal, the TaskContext injection block, TaskBridge.
  Keep UIStateContext separate. Fixes the latent tearing bug. Behavior otherwise
  identical (all SSE wiring carried over verbatim).
- Delete the dead taskProgress lane (field + TaskProgress interface across types/
  setTaskState/commit/ChatService/StorageService/useWidget/MessengerShell/ChatView
  + the permanently-false "Task Progress (N steps)" render block) — nothing ever
  populated it; superseded by per-message ChatMessage.parts.
- Merge TabBar into ShellTabBar (drop the string→WidgetView cast), delete
  blocks/TabBar.tsx; delete the zero-caller useWidget.setState + the resetState
  alias.

type-check clean; 225 vitest pass; lint 0; build ok. Net -184 LOC.
