+++ src/context/WidgetProviders.tsx
@@ -1,28 +1,27 @@
 /**
- * WidgetProviders — composition root that wires UIStateProvider, ChatProvider,
- * and TaskProvider together.
+ * WidgetProviders — composition root that wires UIStateProvider and
+ * ConversationProvider together.
  *
- * Initialization (SSE connection, ChatService hydration) and persistence both
- * live in inner bridge components that have access to all three contexts.
- * React context is the single source of truth; ChatService is the load/persist
- * boundary only, driven by one effect here.
+ * Conversation state ({ messages, task }) lives in ONE store (ConversationProvider);
+ * UI state (open/minimized/mode/loading/error) lives in UIStateProvider. Both
+ * one-time initialization (SSE connection, ChatService hydration) and persistence
+ * live in inner bridge components that can read every context. React context is
+ * the single source of truth; ChatService is the load/persist boundary only,
+ * driven by one effect here.
  */
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect } from 'react';
 
 import { chatService } from '../services/ChatService';
 import { configManager } from '../services/ConfigManager';
 import { sessionManager } from '../services/SessionManager';
 import { StreamClient } from '../services/StreamClient';
-import type { ChatMessage } from '../types';
-import { ChatProvider, useChatContext, useChatContextInternal } from './ChatContext';
-import { TaskProvider, useTaskContext } from './TaskContext';
+import { ConversationProvider, useConversationContext } from './ConversationContext';
 import { UIStateProvider, useUIStateContext } from './UIStateContext';
 
 // Persistence bridge — single context → storage effect (innermost, sees all state).
 const PersistBridge: React.FC<{ previewMode: boolean }> = ({ previewMode }) => {
   const { uiState } = useUIStateContext();
-  const { chatState } = useChatContext();
-  const { taskState } = useTaskContext();
+  const { chatState, taskState } = useConversationContext();
 
   useEffect(() => {
     if (previewMode) return;
@@ -30,7 +29,6 @@ const PersistBridge: React.FC<{ previewMode: boolean }> = ({ previewMode }) => {
       messages: chatState.messages,
       isTaskRunning: taskState.isTaskRunning,
       activeTaskId: taskState.activeTaskId,
-      taskProgress: taskState.taskProgress,
       currentMode: uiState.currentMode,
       isOpen: uiState.isOpen,
       isMinimized: uiState.isMinimized,
@@ -41,7 +39,6 @@ const PersistBridge: React.FC<{ previewMode: boolean }> = ({ previewMode }) => {
     chatState.messages,
     taskState.isTaskRunning,
     taskState.activeTaskId,
-    taskState.taskProgress,
     uiState.currentMode,
     uiState.isOpen,
     uiState.isMinimized,
@@ -51,20 +48,11 @@ const PersistBridge: React.FC<{ previewMode: boolean }> = ({ previewMode }) => {
   return null;
 };
 
-// Inner bridge — has access to UIState and Chat, mounts TaskProvider.
-interface BridgeProps {
-  children: React.ReactNode;
-  previewMode: boolean;
-}
-
-const TaskBridge: React.FC<BridgeProps> = ({ children, previewMode }) => {
-  const { uiState, uiActions } = useUIStateContext();
-  const { _setMessages } = useChatContextInternal();
-
-  // Stable ref so TaskProvider's stopTask never stale-closes over activeTaskId
-  const activeTaskIdRef = useRef<string | null>(null);
-  // Live messages snapshot the SSE reducer wiring reads without stale closures
-  const messagesRef = useRef<ChatMessage[]>([]);
+// Initialization bridge — hydrates from ChatService + connects the stream, then
+// seeds the conversation store. Sees both UIState and Conversation contexts.
+const InitBridge: React.FC<{ children: React.ReactNode; previewMode: boolean }> = ({ children, previewMode }) => {
+  const { uiActions } = useUIStateContext();
+  const { chatActions, taskActions } = useConversationContext();
 
   // One-time initialization: hydrate from ChatService + connect stream.
   useEffect(() => {
@@ -90,10 +78,11 @@ const TaskBridge: React.FC<BridgeProps> = ({ children, previewMode }) => {
         isMinimized: snapshot.isMinimized,
       });
 
-      messagesRef.current = snapshot.messages;
-      _setMessages({ messages: snapshot.messages });
-
-      activeTaskIdRef.current = snapshot.isTaskRunning ? snapshot.activeTaskId : null;
+      chatActions.setMessages(snapshot.messages);
+      taskActions.setTaskState({
+        activeTaskId: snapshot.isTaskRunning ? snapshot.activeTaskId : null,
+        isTaskRunning: snapshot.isTaskRunning,
+      });
 
       if (chatId) {
         const streamClient = StreamClient.getInstance();
@@ -117,28 +106,21 @@ const TaskBridge: React.FC<BridgeProps> = ({ children, previewMode }) => {
   }, []);
 
   return (
-    <TaskProvider
-      previewMode={previewMode}
-      currentMode={uiState.currentMode}
-      setMessages={_setMessages}
-      messagesRef={messagesRef}
-      uiActions={uiActions}
-      activeTaskIdRef={activeTaskIdRef}
-    >
+    <>
       {children}
       <PersistBridge previewMode={previewMode} />
-    </TaskProvider>
+    </>
   );
 };
 
-// UIState → Chat bridge (reads UIState to inject into ChatProvider).
-const UIStateBridge: React.FC<BridgeProps> = ({ children, previewMode }) => {
+// UIState → Conversation bridge (reads UIState to inject into ConversationProvider).
+const UIStateBridge: React.FC<{ children: React.ReactNode; previewMode: boolean }> = ({ children, previewMode }) => {
   const { uiState, uiActions } = useUIStateContext();
 
   return (
-    <ChatProvider previewMode={previewMode} uiActions={uiActions} currentMode={uiState.currentMode}>
-      <TaskBridge previewMode={previewMode}>{children}</TaskBridge>
-    </ChatProvider>
+    <ConversationProvider previewMode={previewMode} currentMode={uiState.currentMode} uiActions={uiActions}>
+      <InitBridge previewMode={previewMode}>{children}</InitBridge>
+    </ConversationProvider>
   );
 };
